requirements.txt
ninja
matplotlib
wandb
pillow
numpy
h5py
pyyaml
gdown
pyrender
protobuf
vedo
pymeshlab
easydict
open3d
pykeops
einops
timm
addict
sharedarray
tensorboard
tensorboardx 
yapf
scipy 
plyfile 
termcolor
pyfqmr
opencv-python
torchmetrics


# flash-attn==2.5.9.post1
# torch-points-kernels==0.6.10
# torch-points3d==1.3.0

# https://pytorch-geometric.readthedocs.io/en/2.3.0/notes/installation.html
# windows run:
# pip install torch_scatter torch_sparse torch_cluster torch_spline_conv torch-geometric -f https://data.pyg.org/whl/torch-2.4.0+cu118.html
# linux run:
# pip install pyg_lib torch_scatter torch_sparse torch_cluster torch_spline_conv torch-geometric -f https://data.pyg.org/whl/torch-2.4.0+cu118.html

# pip install torch==2.4.0 torchvision==0.19.0 torchaudio==2.4.0 --index-url https://download.pytorch.org/whl/cu118

# pip install spconv-cu118